Task_2_rag/requirements.txt
PyMuPDF
numpy
sentence-transformers
transformers
langchain
langchain-text-splitters
streamlit
requests
supabase
python-dotenv
